refactor: extracted room tile widget to a separate file
