--- lib/views/widgets/example_widget.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-
-//TODO: Delete this file after adding any file to this directory
-class ExampleWidget extends StatelessWidget {
-  const ExampleWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
